API: Extract superclass from StrictMetricsEvaluator
Extract the strict evaluation logic from StrictMetricsEvaluator into a new
superclass, StrictEvalVisitor. StrictMetricsEvaluator now uses the visitor
superclass and implements methods to access metrics by ID. This is in
preparation for adding a strict stats visitor that consumes ContentStats
directly.
